datapunk/lib/auth/policy/rollback/validation.py
from typing import Dict, List, Optional, TYPE_CHECKING
import structlog
from dataclasses import dataclass
from enum import Enum

from ...core.exceptions import ValidationError
from ...types import ValidationResult

if TYPE_CHECKING:
    from ....monitoring import MetricsClient

logger = structlog.get_logger()

class RollbackRisk(Enum):
    """
    Risk levels for policy rollback operations, categorizing potential impact.
    Used to help administrators make informed decisions about rollback operations.
    
    Risk levels are determined based on:
    - Number of affected resources
    - Security implications
    - Compliance impact
    - Service disruption potential
    """
    LOW = "low"           # Minor changes, low impact
    MEDIUM = "medium"     # Significant changes, moderate impact
    HIGH = "high"         # Major changes, high impact
    CRITICAL = "critical" # Breaking changes, severe impact

@dataclass
class RollbackValidationResult:
    """
    Comprehensive result of a policy rollback validation operation.
    
    Attributes:
        is_valid: Indicates if rollback can proceed safely
        risk_level: Assessed RollbackRisk level for the operation
        breaking_changes: List of changes that could break existing functionality
        warnings: Non-critical issues that should be reviewed
        recommendations: Suggested actions to mitigate risks
    """
    is_valid: bool
    risk_level: RollbackRisk
    breaking_changes: List[str]
    warnings: List[str]
    recommendations: List[str]

class RollbackValidator:
    """
    Validates policy rollbacks for safety and consistency.
    
    This validator ensures that policy rollbacks don't introduce breaking changes,
    security vulnerabilities, or compliance violations. It performs comprehensive
    checks across multiple dimensions including resource access, security implications,
    and compliance requirements.
    
    Dependencies:
        - MetricsClient for operational monitoring
        - structlog for contextual logging
    """
    
    def __init__(self, metrics: 'MetricsClient'):
        """
        Initialize validator with metrics client for operational monitoring.
        
        Args:
            metrics: Client for recording validation metrics and monitoring
        """
        self.metrics = metrics
        self.logger = logger.bind(component="rollback_validator")
    
    async def validate_rollback(self,
                              current_policy: Dict,
                              rollback_policy: Dict,
                              affected_keys: List[str]) -> RollbackValidationResult:
        """
        Validate a policy rollback operation for safety and consistency.
        
        Performs comprehensive validation across multiple dimensions:
        1. Resource access changes
        2. Security implications
        3. Compliance impact
        
        Args:
            current_policy: Currently active policy configuration
            rollback_policy: Target policy configuration to roll back to
            affected_keys: List of policy keys that will be modified
            
        Returns:
            RollbackValidationResult containing validation outcome and details
            
        Raises:
            ValidationError: If validation process encounters unexpected errors
            
        Note:
            - Validation is async to handle potential external service calls
            - Breaking changes automatically invalidate the rollback
            - Risk level is assessed based on cumulative impact
        """
        try:
            breaking_changes = []
            warnings = []
            recommendations = []
            
            # TODO: Consider adding parallel validation for performance optimization
            # FIXME: Add rate limiting for external service calls
            
            self._validate_resource_access(
                current_policy,
                rollback_policy,
                breaking_changes,
                warnings
            )
            
            self._validate_security_changes(
                current_policy,
                rollback_policy,
                breaking_changes,
                warnings,
                recommendations
            )
            
            self._validate_compliance_changes(
                current_policy,
                rollback_policy,
                breaking_changes,
                warnings,
                recommendations
            )
            
            risk_level = self._assess_risk_level(
                breaking_changes,
                warnings,
                affected_keys
            )
            
            return RollbackValidationResult(
                is_valid=len(breaking_changes) == 0,
                risk_level=risk_level,
                breaking_changes=breaking_changes,
                warnings=warnings,
                recommendations=recommendations
            )
            
        except Exception as e:
            # NOTE: Consider adding more detailed error categorization
            self.logger.error("rollback_validation_failed",
                            error=str(e))
            raise ValidationError(f"Rollback validation failed: {str(e)}") 